Fix: Preferences not being retrieved because `__package__` does not match addon directory
Using `__package__` assumes the script is in the top-level of the addon directory. Due to how the VSCode extension for Blender development works, this is not the case for this addon.

To work around this, the preferences lookup uses a hard coded name during development, and the build script replaces that with `__package__` when zipping the release versions of the addon.