Refactor: Use `bl_idname` to identify overriding menus instead of `__name__`
The original class name for the menus being replaced
are inconsistent and less than ideal. Therefore they were
renamed to be easier to understand during development.

However this broke compatibility with other addons that
attach to the original menus, since the bl_idnames don't match.

The decided solution for this is to make it so that the registered
bl_idnames for the menus are based on the original ones they're replacing,
while still keeping the class names in their altered state.